add ダメージ集計の関数を定義
