+++ config/settings/cb_convex.yaml
@@ -11,7 +11,7 @@ DECLARE_MESSAGE_ID:
   a:
     PLAN: '905442795991482449'
     DECLARE: '905442797866348584'
-    STATUS: '905442799183339581'
+    DAMAGE: '905442799183339581'
   b:
     STATUS: '885121309795840010'
     PLAN: '885121311230271508'
