+++ src/client/messageReactionRemove.ts
@@ -1,7 +1,6 @@
 import * as Discord from 'discord.js'
 import Option from 'type-of-option'
 import ThrowEnv from 'throw-env'
-import * as declare from '../convex/declare/react'
 
 /**
  * リアクションのイベントに応じて適切な処理を実行する
@@ -17,7 +16,5 @@ export const MessageReactionRemove = async (
 
   let comment: Option<string>
   user = user as Discord.User
-
-  comment = await declare.NoticeCancel(react, user)
-  if (comment) return console.log(comment)
+  comment
 }
