+++ src/client/messageUpdate.ts
@@ -1,7 +1,6 @@
 import * as Discord from 'discord.js'
 import Option from 'type-of-option'
 import ThrowEnv from 'throw-env'
-import * as declare from '../convex/declare/status'
 import * as edit from '../convex/plan/edit'
 
 /**
@@ -14,9 +13,6 @@ export const MessageUpdate = async (msg: Discord.Message | Discord.PartialMessag
 
   let comment: Option<string>
   msg = msg as Discord.Message
-
-  comment = await declare.Edit(msg)
-  if (comment) return console.log(comment)
 
   comment = await edit.Message(msg)
   if (comment) return console.log(comment)
